refactor: Updating Migration Change int to Biting

+++ migration/url-table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE url(
-  id INT primary key,
+  id bigint primary key,
   long_url TEXT not null,
   short_url TEXT not null
 )
